E2E tests and scripts to run them.
This is the start of the basic test structure for e2e.
Not sure if this is the best structure and tools (jest) for e2e
but for now since there is only one test this will do.

Later on I will want to remove the hard 3 sec sleep. Doing this
for know to finish out the task.